Refactor MAPElites package for correctness and readability
- Change feature vector possibility count to double to prevent overflow.
- Extract duplicated TestChromosome utility methods to TestChromosomeUtils.
- Clean up dead code and improve code style in MAPElites.
- Add tests for FeatureVector and TestChromosomeUtils.

+++ client/src/main/java/org/evosuite/ga/metaheuristics/mapelites/TestResultObserver.java
@@ -51,7 +51,7 @@ public class TestResultObserver extends ExecutionObserver implements Serializabl
         Arrays.sort(this.inspectors, (a, b) -> a.getMethodCall().compareTo(b.getMethodCall()));
     }
 
-    public int getPossibilityCount() {
+    public double getPossibilityCount() {
         return FeatureVector.getPossibilityCount(this.inspectors);
     }
